+++ RiskOfChaos/Config/UI.cs
@@ -1,6 +1,10 @@
 ﻿using BepInEx.Configuration;
 using RiskOfChaos.ConfigHandling;
+using RiskOfChaos.Trackers;
 using RiskOfOptions.OptionConfigs;
+using RoR2;
+using RoR2.UI;
+using System;
 using UnityEngine;
 
 namespace RiskOfChaos
@@ -35,6 +39,36 @@ namespace RiskOfChaos
                                    .OptionConfig(new CheckBoxConfig())
                                    .Build();
 
+            public enum NextEffectTimerDisplayType : byte
+            {
+                Never,
+                WhenRunTimerUnavailable,
+                Always
+            }
+
+            public static readonly ConfigHolder<NextEffectTimerDisplayType> NextEffectTimerDisplayMode =
+                ConfigFactory<NextEffectTimerDisplayType>.CreateConfig("Next Effect Timer Display Mode", NextEffectTimerDisplayType.WhenRunTimerUnavailable)
+                                                         .Description($"Displays how much time is left until the next effect.\n\n{NextEffectTimerDisplayType.Never}: The time remaining is never displayed.\n{NextEffectTimerDisplayType.WhenRunTimerUnavailable}: Displays time remaining only when the regular run timer is paused or otherwise not visible.\n{NextEffectTimerDisplayType.Always}: Time remaining is always displayed")
+                                                         .OptionConfig(new ChoiceConfig())
+                                                         .ValueValidator(CommonValueValidators.DefinedEnumValue<NextEffectTimerDisplayType>())
+                                                         .Build();
+
+            public static bool ShouldShowNextEffectTimer(HUD hud)
+            {
+                switch (NextEffectTimerDisplayMode.Value)
+                {
+                    case NextEffectTimerDisplayType.Never:
+                        return false;
+                    case NextEffectTimerDisplayType.WhenRunTimerUnavailable:
+                        Run run = Run.instance;
+                        return run && ((run.isRunStopwatchPaused && General.RunEffectsTimerWhileRunTimerPaused.Value) || !RunTimerUITracker.IsAnyTimerVisibleForHUD(hud));
+                    case NextEffectTimerDisplayType.Always:
+                        return true;
+                    default:
+                        throw new NotImplementedException();
+                }
+            }
+
             internal static void Bind(ConfigFile file)
             {
                 void bindConfig<T>(ConfigHolder<T> config)
@@ -47,6 +81,8 @@ namespace RiskOfChaos
                 bindConfig(ActiveEffectsTextColor);
 
                 bindConfig(DisplayNextEffect);
+
+                bindConfig(NextEffectTimerDisplayMode);
             }
         }
     }