+++ RiskOfChaos/UI/NextEffectDisplay/EffectDisplayData.cs
@@ -3,8 +3,5 @@ using RiskOfChaos.EffectHandling.Formatting;
 
 namespace RiskOfChaos.UI.NextEffectDisplay
 {
-    public readonly record struct EffectDisplayData(ChaosEffectIndex EffectIndex, float TimeRemaining, EffectNameFormatter NameFormatter)
-    {
-        public static readonly EffectDisplayData None = new EffectDisplayData(ChaosEffectIndex.Invalid, -1f, null);
-    }
+    public readonly record struct EffectDisplayData(ChaosEffectIndex EffectIndex, float TimeRemaining, EffectNameFormatter NameFormatter);
 }