+++ RiskOfChaos/UI/NextEffectDisplay/NextEffectDisplayController.cs
@@ -1,4 +1,6 @@
-﻿using RiskOfChaos.EffectHandling;
+﻿using RiskOfChaos.EffectDefinitions;
+using RiskOfChaos.EffectHandling;
+using RiskOfChaos.EffectHandling.Controllers;
 using RoR2;
 using RoR2.UI;
 using UnityEngine;
@@ -12,6 +14,7 @@ namespace RiskOfChaos.UI.NextEffectDisplay
         public Image BackdropImage;
         public AnimateUIAlpha FlashController;
 
+        float _lastDisplayedTimeRemaining;
         ChaosEffectIndex _currentDisplayingEffectIndex;
 
         void Awake()
@@ -21,25 +24,41 @@ namespace RiskOfChaos.UI.NextEffectDisplay
             EffectText.rectTransform.anchoredPosition = Vector2.zero;
         }
 
+        void OnEnable()
+        {
+            beginFlash();
+        }
+
+        void beginFlash()
+        {
+            if (FlashController)
+            {
+                FlashController.time = 0f;
+            }
+        }
+
         public void DisplayEffect(EffectDisplayData displayData)
         {
-            ChaosEffectInfo effectInfo = ChaosEffectCatalog.GetEffectInfo(displayData.EffectIndex);
-
-            string effectName = effectInfo?.GetDisplayName(displayData.NameFormatter, EffectNameFormatFlags.RuntimeFormatArgs) ?? "NULL";
-
             string timeRemainingString = displayData.TimeRemaining.ToString(displayData.TimeRemaining >= 10f ? "F0" : "F1");
 
-            EffectText.text = Language.GetStringFormatted("CHAOS_NEXT_EFFECT_DISPLAY_FORMAT", effectName, timeRemainingString);
-
-            if (_currentDisplayingEffectIndex != displayData.EffectIndex)
+            if (displayData.EffectIndex != ChaosEffectIndex.Invalid)
             {
-                if (FlashController)
-                {
-                    FlashController.time = 0f;
-                }
+                ChaosEffectInfo effectInfo = ChaosEffectCatalog.GetEffectInfo(displayData.EffectIndex);
+                string effectName = effectInfo?.GetDisplayName(displayData.NameFormatter, EffectNameFormatFlags.RuntimeFormatArgs) ?? "NULL";
+                EffectText.text = Language.GetStringFormatted("CHAOS_NEXT_EFFECT_DISPLAY_FORMAT", effectName, timeRemainingString);
+            }
+            else
+            {
+                EffectText.text = Language.GetStringFormatted("CHAOS_NEXT_EFFECT_TIME_REMAINING_DISPLAY_FORMAT", timeRemainingString);
+            }
+
+            if (_currentDisplayingEffectIndex != displayData.EffectIndex || displayData.TimeRemaining - _lastDisplayedTimeRemaining > 1f)
+            {
+                beginFlash();
             }
 
             _currentDisplayingEffectIndex = displayData.EffectIndex;
+            _lastDisplayedTimeRemaining = displayData.TimeRemaining;
         }
     }
 }